requirements.txt
acme==0.19.0
#apt-xapian-index==0.45
asn1crypto==0.22.0
beautifulsoup4==4.5.1
boto==2.48.0
CEPBR==0.1.0
certbot==0.19.0
certbot-nginx==0.19.0
certifi==2017.4.17
chardet==3.0.4
Cheetah==2.4.4
#cloud-init==0.7.5all
colorama==0.2.5
ConfigArgParse==0.11.0
configobj==4.7.2
#cryptography==1.9
Django==1.8.13
django-allauth==0.32.0
django-appconf==1.0.2
django-cache-manager==0.5
django-cacheops==3.1.1
django-carton==1.2.1
django-compressor==2.1
django-contrib-comments==1.7.3
django-dajax==0.9.2
django-dajaxice==0.7
django-debug-toolbar==1.6
django-extensions==1.7.4
django-filer==1.2.5
django-htmlmin==0.10.0
#django-localflavor-br==1.0
django-media-manager==3.4.2
django-mptt==0.8.6
django-polymorphic==0.8.1
#django-redis-cache==1.7.1
django-registration-redux==1.4
django-robots==2.0
Django-Select2==5.8.9
django-ses==0.8.1
django-smart-selects==1.2.9
django-social-auth==0.7.28
django-suit==0.2.21
django-suit-ckeditor==0.0.2
django-threadedcomments==1.0.1
django-widget-tweaks==1.4.1
django-wysiwyg-redactor==0.4.9.1
dnspython==1.15.0
easy-thumbnails==2.3
enum34==0.9.23
funcy==1.7.2
gevent==1.0
greenlet==0.4.2
gunicorn==17.5
html2text==2016.9.19
html5lib==0.999
httplib2==0.10.3
idna==2.5
ipaddress==1.0.17
jsonpatch==1.3
jsonpointer==1.0
#Landscape-Client==14.12
mock==1.0.1
#MySQL-python==1.2.5
ndg-httpsclient==0.4.2
oauth==1.0.1
oauth2==1.9.0.post1
oauthlib==2.0.2
#PAM==0.4.2
parsedatetime==2.1
Pillow==3.4.2
prettytable==0.7.2
psutil==5.0.1
psycopg2
pusher==1.5
pyasn1==0.1.9
pycrypto==2.6.1
PyICU
pyOpenSSL==17.3.0
pyparsing==2.0.1
PyPDF2==1.26.0
pyRFC3339==1.0
pysendy==0.0.5
pyserial==2.6
#python-apt===0.9.3.5ubuntu2
python-dateutil==2.5.3
#python-debian===0.1.21-nmu2ubuntu2
python-openid==2.2.5
python-redis-lock==3.2.0
python2-pythondialog==3.3.0
pytz==2016.7
PyYAML==3.10
rcssmin==1.0.6
redis==2.10.5
#redis-cache==0.1.3
#redis-structures==0.1.6
reportlab==3.3.0
requests==2.18.1
requests-oauthlib==0.8.0
rjsmin==1.0.12
simplejson==3.8.2
six==1.11.0
sorl-thumbnail==12.3
sqlparse==0.2.1
ssh-import-id==3.21
#Twisted-Core==13.2.0
Unidecode==0.4.19
#uniqid==0.0.2https://github.com/Riamse/python-uniqid
urllib3==1.21.1
#virtualenv==1.11.4
#vital-tools==0.1.7
xhtml2pdf==0.0.6
xmltodict==0.10.2
zope.component==4.0.2
zope.event==4.0.1
zope.hookable==3.4.1
zope.interface==4.0.5
